some fixes from int to long

+++ src/main/java/com/synectiks/cms/repository/AddressRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "address", path = "address")
-public interface AddressRepository extends PagingAndSortingRepository<Address, Integer> {
+public interface AddressRepository extends PagingAndSortingRepository<Address, Long> {
 }
