+++ src/main/java/com/synectiks/cms/repository/AuthorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "author", path = "author")
-public interface AuthorRepository extends PagingAndSortingRepository<Author, Integer> {
+public interface AuthorRepository extends PagingAndSortingRepository<Author, Long> {
 }
